requirements.txt
black==24.3.0
python-dotenv==1.0.1
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
marshmallow==3.21.1
celery==5.3.6
redis==5.0.3